feature: create new light
