+++ src/main.jsx
@@ -5,8 +5,8 @@ import App from './App.jsx'
 import './index.css'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
-  <React.StrictMode>
+  <>
     <Leva />
     <App />
-  </React.StrictMode>
+  </>
 )
